7 script para jugar los cuentos

+++ database/migrations/2024_09_14_055722_create_puntajes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('aciertos');
             $table->integer('fallas');
             $table->integer('total_palabras');
-            $table->integer('puntaje');
+            $table->float('puntaje');
             $table->timestamps();
         });
     }
